feat: Refactor homepage components and introduce File Operations context
- Removed legacy Toolbar component and updated imports.
- Added DialogsContainer component to manage various dialog states.
- Introduced EmptyState and ErrorMessage components for better user feedback.
- Implemented FileManager component to handle file operations and UI interactions.
- Created FileOperationsProvider for managing dialog states and context.
- Added Header component for navigation and stats display.
- Included LoadingSpinner component for loading states.
- Updated index file to export new components for easier imports.
