--- components/Toolbar.tsx
@@ -1,3 +1,0 @@
-"use client";
-// This is the legacy file - please use components/Toolbar/index.tsx instead
-export { default } from "./Toolbar/index";
